requirements.txt
matplotlib==3.3.4
numpy==1.20.0
pandas==1.2.1
Pillow==8.1.0